Extract some related methods to CheckoutSteps concern
